requirements.txt
google-generativeai
llama-index-llms-gemini
llama-index-embeddings-gemini
python-dotenv